feat: code refactroing of login
