+++ react-masterclass/package.json
@@ -6,6 +6,7 @@
     "@apollo/react-hooks": "3.0.0",
     "@emotion/core": "10.0.21",
     "@emotion/styled": "10.0.17",
+    "@testing-library/jest-dom": "4.1.0",
     "@reach/router": "1.2.1",
     "@rebass/forms": "4.0.6",
     "@rebass/layout": "4.0.6",
@@ -76,5 +77,33 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
+  },
+  "jest": {
+    "collectCoverageFrom": [
+      "src/**/*.js",
+      "!src/index.js",
+      "!src/stories/**/*.js",
+      "!src/utils.js",
+      "!src/timers/**/*.js",
+      "!src/basic-mock/**/*.js"
+    ],
+    "coverageThreshold": {
+      "global": {
+        "branches": 100,
+        "functions": 100,
+        "lines": 100,
+        "statements": 100
+      }
+    }
+  },
+  "devDependencies": {
+    "cross-env": "^6.0.2",
+    "enzyme": "3.10.0",
+    "enzyme-adapter-react-16": "1.14.0",
+    "husky": "^3.0.8",
+    "jest-enzyme": "7.1.1",
+    "lint-staged": "^9.4.1",
+    "node-sass": "4.12.0",
+    "prettier": "^1.18.2"
   }
 }
