+++ react-masterclass/src/Home.jsx
@@ -4,7 +4,7 @@ import gql from 'graphql-tag.macro';
 import { Box } from 'rebass';
 
 import Pages from './pages/index';
-import Login from './pages/login/Login';
+import Login from './Login';
 import Navbar from './components/common/Navbar';
 
 export const AUTHENTICATED_QUERY = gql`
